Big refactor: Uint for edges, define max width. To use a much bigger network I needed to do some mayor changes: - Given the longest street is ~16km, I could not create a map numLanes*maxWidth but do an overflow of a fix number, this changes the logic to compute the byte in memory - Given the size of the number of edges, also change the logic in the intersections - The following changes made changes in the edge, intersection, and person data.

+++ LivingCity/traffic/cudaTrafficPerson.h
@@ -55,8 +55,8 @@ struct CUDATrafficPerson {
 
   //////////////////////////
   // current edge (from edgeData)
-  unsigned short edgeNumLanes;//total number of lanes in that edge
-  unsigned short edgeNextInters;
+  unsigned short edgeNumLanes;  //total number of lanes in that edge
+  unsigned int edgeNextInters;
   float length;
   float maxSpeedMperSec;
   /////////////////////////
@@ -69,9 +69,8 @@ struct CUDATrafficPerson {
   ///////////////////////////
 
 
-  unsigned short currPathEdge;
-  unsigned short
-  personPath[80];//change CUDATrafficPersonShortestPath::calculateSeveralPeopleRoute
+  unsigned int currPathEdge;
+  unsigned int personPath[80];//change CUDATrafficPersonShortestPath::calculateSeveralPeopleRoute
   // data
   unsigned short num_steps;
   unsigned int color;
